hago que delete borre files, empiezo a ordenar folders y categorias en proceso...

+++ src/controllers/admin/productsController.js
@@ -1,4 +1,6 @@
 const fs = require("fs/promises");
+const {existsSync} = require('fs')
+
 const sharp = require("sharp");
 const path = require("path");
 const { validationResult } = require("express-validator");
@@ -11,14 +13,14 @@ const getAdminView = async (get, res) => {
         const idProductos = productos.map(producto => producto.id);
         const nombresProductos = productos.map(producto => producto.product_name);
         const skuProductos = productos.map(producto => producto.sku);
-        res.render("admin", { nombresProductos, skuProductos, idProductos });   
+        res.render("./admin/productos/admin", { nombresProductos, skuProductos, idProductos });   
     } catch (error) {
         console.log(error)
     }
 }
 
 const getCreateProductView = (req, res) => {
-    res.render("create");
+    res.render("./admin/productos/create");
 }
 
 const createProduct = async (req, res) => {
@@ -26,7 +28,7 @@ const createProduct = async (req, res) => {
     const errors = validationResult(req)
 
     if (!errors.isEmpty()) {
-        return res.render("create", {
+        return res.render("./admin/productos/create", {
             values: req.body,
             errors: errors.array(),
         });
@@ -62,7 +64,7 @@ const getEditProductView = async (req, res) => {
     try {
         const producto = await model.findByPk(req.params.id);
         if (producto) {
-            res.render("edit", {producto});  
+            res.render("./admin/productos/edit", {producto});  
         } else {
             res.status(404).send("El producto solicitado no existe");
         }
@@ -81,7 +83,7 @@ const editProduct = async (req, res) => {
 
     if (!errors.isEmpty()) {
         const producto = await model.findByPk(req.params.id)
-        return res.render("./edit", {
+        return res.render("./admin/productos/edit", {
             producto: producto,
             values: req.body,
             errors: errors.array(),
@@ -123,35 +125,37 @@ const editProduct = async (req, res) => {
 }
 
 const deleteProduct = async (req, res) => {
+    
     try {
         const deleted = await model.destroy({
             where: {
                 id: req.params.id
             }
         })
-
-        if (deleted == 1){
-            console.log("Eliminando producto");
-
-
-            await fs.unlink(
-                path.resolve(
-                    __dirname,`../../../public/uploads/${req.params.id}-1.webp`
-                )
-            );
-            await fs.unlink(
-                path.resolve(
-                    __dirname,`../../../public/uploads/${req.params.id}-box.webp`
-                )
-            );
-        }
         
+        if (deleted == 1){
+            if (existsSync(path.resolve(__dirname,`../../../public/uploads/${req.params.id}-1.webp`))){
+                await fs.unlink(
+                    path.resolve(
+                        __dirname,`../../../public/uploads/${req.params.id}-1.webp`
+                        )
+                        );
+                    }
+
+            if (existsSync(path.resolve(__dirname,`../../../public/uploads/${req.params.id}-box.webp`))){
+                await fs.unlink(
+                    path.resolve(
+                        __dirname,`../../../public/uploads/${req.params.id}-box.webp`
+                        )
+                        );
+                    }
+        }
         res.redirect("/admin/");
     } catch (error) {
         console.log(error);
     }
-}
-
+};
+                    
 module.exports = {
     getCreateProductView,
     createProduct,
